pyproject.toml
[tool.poetry]
name = "langchain-expression-language-test"
version = "0.1.0"
description = "Test LCEL"
authors = ["Sebastian Montero <sebastian.montero@gmail.com>"]
readme = "README.md"
packages = [{include = "langchain_expression_language_test"}]

[tool.poetry.dependencies]
python = ">=3.8.1,<4.0"
python-dotenv = "^1.0.1"
langchain-core = "^0.1.44"
langchain-community = "^0.0.33"
langchain-openai = "^0.1.3"
langchain = {extras = ["docarray"], version = "^0.1.16"}
docarray = "0.32.0"
pydantic = "1.10.9"
asyncio = "^3.4.3"
cohere = "^5.3.2"
langchain-cohere = "^0.1.3"


[tool.poetry.group.dev.dependencies]
autopep8 = "^2.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
